libs/skyline-apiserver/pyproject.toml
[tool.poetry]
name = "skyline-apiserver"
version = "0.1.0"
description = ""
license = "Apache-2.0"
authors = ["OpenStack <openstack-discuss@lists.openstack.org>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "*"}
PyYAML = "*"
immutables = "*"
orjson = "*"
ujson = "*"
uvicorn = {extras = ["standard"], version = "^0.12.1"}
gunicorn = "*"
python-keystoneclient = "3.21.*"
python-cinderclient = "5.0.*"
python-glanceclient = "2.17.*"
python-heatclient = "1.18.*"
python-neutronclient = "6.14.*"
python-novaclient = "15.1.*"
python-octaviaclient = "1.10.*"
osc-placement = "1.7.*"
keystoneauth1 = "3.17.*"
email-validator = "^1.1.1"
python-jose = "^3.2.0"
passlib = "^1.7.2"
alembic = "^1.4.2"
bcrypt = "^3.2.0"
hiyapyco = "^0.4.16"
httpx = "^0.16.1"
sqlalchemy = "1.3.*"
databases = "*"
aiomysql = "^0.0.21"
pymysql = "*"
skyline-policy-manager = "*"
skyline-log = "*"
skyline-config = "*"

[tool.poetry.dev-dependencies]
isort = "*"
black = "^21.5b1"
add-trailing-comma = "*"
flake8 = "*"
mypy = "*"
types-PyYAML = "*"
pytest = "*"
pytest-xdist = {extras = ["psutil"], version = "*"}
aiosqlite = "*"
asgi-lifespan = "*"
pytest-asyncio = "*"
skyline-policy-manager = {path = "../skyline-policy-manager", develop = true}
skyline-log = {path = "../skyline-log", develop = true}
skyline-config = {path = "../skyline-config", develop = true}

[tool.poetry.scripts]
swagger-generator = 'skyline_apiserver.cmd.generate_swagger:main'
config-sample-generator = 'skyline_apiserver.cmd.generate_sample_config:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
